Extract BaseFileSegment & BracketedSegment
