+++ src/sqlfluff/core/parser/segments/__init__.py
@@ -2,32 +2,32 @@
 
 from sqlfluff.core.parser.segments.base import (
     BaseSegment,
-    BaseFileSegment,
-    UnparsableSegment,
-    BracketedSegment,
     FixPatch,
     SourceFix,
+    UnparsableSegment,
 )
-from sqlfluff.core.parser.segments.generator import SegmentGenerator
-from sqlfluff.core.parser.segments.raw import (
-    RawSegment,
-    CodeSegment,
-    UnlexableSegment,
-    CommentSegment,
-    WhitespaceSegment,
-    NewlineSegment,
-    KeywordSegment,
-    SymbolSegment,
-)
+from sqlfluff.core.parser.segments.bracketed import BracketedSegment
 from sqlfluff.core.parser.segments.ephemeral import EphemeralSegment, allow_ephemeral
+from sqlfluff.core.parser.segments.file import BaseFileSegment
+from sqlfluff.core.parser.segments.generator import SegmentGenerator
 from sqlfluff.core.parser.segments.meta import (
-    MetaSegment,
-    Indent,
     Dedent,
-    ImplicitIndent,
-    TemplateSegment,
     EndOfFile,
+    ImplicitIndent,
+    Indent,
+    MetaSegment,
     TemplateLoop,
+    TemplateSegment,
+)
+from sqlfluff.core.parser.segments.raw import (
+    CodeSegment,
+    CommentSegment,
+    KeywordSegment,
+    NewlineSegment,
+    RawSegment,
+    SymbolSegment,
+    UnlexableSegment,
+    WhitespaceSegment,
 )
 
 __all__ = (
